backend/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
alembic==1.12.1
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# File handling
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0

# HTTP client
httpx==0.25.2

# AI/LLM
openai==1.6.1

# Utilities
pyyaml==6.0.1
structlog==24.1.0
numpy==1.24.3